Move runtime meta data into code
